Add analytics endpoints, queries and pagination
Introduce analytics API surface and backend support: new AnalyticsController with endpoints for top host performance, monthly dynamics, market price percentiles, apartment occupancy (with date range + paging) and host portfolio (paged) including basic input validation. Add MediatR queries and handlers for each analytics use case. Update DTOs to use long for booking/count fields. Extend IAnalyticsSqlRepository and implement pagination in AnalyticsSqlRepository; wire LIMIT/OFFSET parameters into SQL queries (GetApartmentOccupancy.sql, GetHostsPortfolio.sql). These changes enable paged analytics results and align numeric types for large counts.

+++ BookingSystemWeb/BookingSystem.Application/Common/DTOs/ApartmentOccupancyDto.cs
@@ -3,6 +3,6 @@
 public record ApartmentOccupancyDto(
     Guid ApartmentId,
     string Title,
-    int BookingsCount,
-    int TotalBookedDays
+    long BookingsCount,
+    long TotalBookedDays
 );
